Fix a few issues related to trivia generation for add/remove suppression fixes from the error list. Unit tests have been added for the same.
